+++ hw/analog.h
@@ -9,7 +9,7 @@
 void ANALOG_Configuration(void);
 void ANALOG_StartDMA(void);
 bool ANALOG_IsReady(void);
-uint16_t* ANALOG_GetFullBuff(void);
+uint8_t* ANALOG_GetFullBuff(void);
 uint16_t ANALOG_GetRate(void);
 
 #endif
